benchmark encoding (capnp / json)
serde_json is very competitive, actually faster until we get above
100B.  It's only on the longest messages, 10 kB as a worst-case
scenario, that capnp really pulls ahead.  At 10k RPS, 4μs on encoding
would be fine (4% of the time budget); it's only if we want to do 100k
RPS *and* a typical request is 10 kB that we need to keep encoding
speed < 1μs.

I think that means JSON would be entirely viable, and likely the
better choice given familiarity, and that we need it anyway for
Sidekiq jobs.